- remove the SG_GLxxxx_H #defines, since OSG provides its own versions  - this exposed a bizarre issue on Mac where dragging in <AGL/agl.h> in    extensions.hxx was pulling in all of Carbon to the global namespace    - very scary. As a result, I now need to explicitly include CoreFoundation    in fg_init.cxx.  - change SG_USING_STD(x) to using std::x
Issues:

 - the logic for X11 and Win32 in RenderTexture and extensions is tortured,
   please see if you agree I got all the ifdefs correct.

+++ simgear/screen/RenderTexture.cpp
@@ -496,7 +496,7 @@ bool RenderTexture::Initialize(int width, int height,
         fprintf(stderr, "Couldn't get current context!");
    
         CGLPixelFormatObj pixFormat = NULL;
-        long int iNumFormats;
+        GLint iNumFormats;
         CGLError error;
 
         // Copy the _pixelFormatAttribs into another array to fix
@@ -553,7 +553,7 @@ bool RenderTexture::Initialize(int width, int height,
            return false;
        }
 
-       long screen;
+       GLint screen;
        if (error = CGLGetVirtualScreen(CGLGetCurrentContext(), &screen))
        {
            _cglCheckError(error);
